requirements.txt
streamlit==1.31.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
python-docx==1.0.1
wheel==0.42.0